refactor: useBoolean hook
