+++ src/hooks/index.ts
@@ -3,7 +3,7 @@ export { default as useHistory } from './useHistory';
 export { default as useUpdates } from './useUpdates';
 export { default as useFullscreenMode } from './useFullscreenMode';
 export { default as useNovelTrackerInfo } from './useNovelTrackerInfo';
-export { default as useBoolean } from './useBoolean';
+export { default as useBoolean } from './common/useBoolean';
 export { useTheme } from './useTheme';
 export { default as useCategories } from './useCategories';
 export { useCategorySettings, useLibrarySettings } from './useSettings';
