+++ src/screens/BrowseSourceScreen/components/FilterBottomSheet.tsx
@@ -14,7 +14,7 @@ import { FlatList } from 'react-native-gesture-handler';
 import { Button, IconButtonV2 } from '@components/index';
 import { Checkbox } from '@components/Checkbox/Checkbox';
 import { Picker } from '@react-native-picker/picker';
-import useBoolean from '@hooks/useBoolean';
+import { useBoolean } from '@hooks';
 import { overlay } from 'react-native-paper';
 import { getValueFor } from './filterUtils';
