+++ src/screens/settings/SettingsBackupScreen/index.tsx
@@ -5,8 +5,7 @@ import { ScreenContainer } from '@components/Common';
 import { useTheme } from '@hooks/useTheme';
 import { Appbar, List } from '@components';
 import { Portal } from 'react-native-paper';
-import useBoolean from '@hooks/useBoolean';
-// import { remoteBackup, remoteRestore } from '@services/backup/remote';
+import { useBoolean } from '@hooks';
 import { BackupSettingsScreenProps } from '@navigators/types';
 import GoogleDriveModal from './Components/GoogleDriveModal';
 import { useMMKVBoolean } from 'react-native-mmkv';
